# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nastech TTS Agent Gateway",
    "description": "A structured expressive-speech gateway that compiles NastechML to Fish S2 controls. Use /v1/agent/compile before synthesis when an agent needs auditable provider intent.",
    "version": "0.3.0"
  },
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_v1_capabilities_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Capabilities V1 Capabilities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/tools": {
      "get": {
        "summary": "Agent Tools",
        "operationId": "agent_tools_v1_agent_tools_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentToolDescriptor"
                  },
                  "title": "Response Agent Tools V1 Agent Tools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/compile": {
      "post": {
        "summary": "Compile Speech",
        "operationId": "compile_speech_v1_agent_compile_post",
        "parameters": [
          {
            "name": "traceparent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Traceparent"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compile Speech V1 Agent Compile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/speech": {
      "post": {
        "summary": "Synthesize Speech",
        "operationId": "synthesize_speech_v1_agent_speech_post",
        "parameters": [
          {
            "name": "traceparent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Traceparent"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSpeechRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audio/speech": {
      "post": {
        "summary": "Openai Compatible Speech",
        "operationId": "openai_compatible_speech_v1_audio_speech_post",
        "parameters": [
          {
            "name": "traceparent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Traceparent"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAISpeechRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentCompileRequest": {
        "properties": {
          "markup": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Markup"
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id"
          },
          "output_format": {
            "type": "string",
            "pattern": "^(wav|mp3|opus|pcm)$",
            "title": "Output Format",
            "default": "wav"
          },
          "sample_rate": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 48000.0,
                "minimum": 8000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate",
            "default": 44100
          },
          "latency": {
            "type": "string",
            "pattern": "^(low|balanced|normal)$",
            "title": "Latency",
            "default": "normal"
          },
          "temperature": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          }
        },
        "type": "object",
        "required": [
          "markup"
        ],
        "title": "AgentCompileRequest",
        "description": "A structured agent request expressed in stable NastechML."
      },
      "AgentSpeechRequest": {
        "properties": {
          "markup": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Markup"
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id"
          },
          "output_format": {
            "type": "string",
            "pattern": "^(wav|mp3|opus|pcm)$",
            "title": "Output Format",
            "default": "wav"
          },
          "sample_rate": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 48000.0,
                "minimum": 8000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate",
            "default": 44100
          },
          "latency": {
            "type": "string",
            "pattern": "^(low|balanced|normal)$",
            "title": "Latency",
            "default": "normal"
          },
          "temperature": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          }
        },
        "type": "object",
        "required": [
          "markup"
        ],
        "title": "AgentSpeechRequest",
        "description": "NastechML request that produces a provider audio response."
      },
      "AgentToolDescriptor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "input_schema"
        ],
        "title": "AgentToolDescriptor"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OpenAISpeechRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "nastech-fish-s2"
          },
          "input": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Input"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "response_format": {
            "type": "string",
            "pattern": "^(wav|mp3|opus|pcm)$",
            "title": "Response Format",
            "default": "wav"
          },
          "speed": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.5,
            "title": "Speed",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "OpenAISpeechRequest",
        "description": "Small OpenAI-compatible request shape for existing agent clients."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}